Move RepositoryPaymentManager.cs
