+++ Persistence/Persistence.csproj
@@ -19,7 +19,6 @@
     <Folder Include="Repositories\HR\" />
     <Folder Include="Repositories\Partner\" />
     <Folder Include="Repositories\Master\" />
-    <Folder Include="Repositories\Payment\" />
     <Folder Include="Repositories\Users\" />
   </ItemGroup>
